removed the category filtering from yesterday to try a new angle
